Fixed bug due to extra source directive
There was a bug caused by the `.bashrc` source directive at the
start of the snakefile

On a clean Ubuntu 18.04 container envs were being activated for each
rule but then seemingly deactivated when snakemake started to execute
the shell component.

This caused the python2.7 reliant phylophlan and checkm test scripts to
error.

For example, output from adding `conda env list` to the `test_checkm`
rule shows the correct activation before the shell but the wrong activated
environment (shown by asterisk) in the shell instance for the rule:

```
activating conda environment: /home/root/MAGpy/.snakemake/conda/377a289b
           /home/root/MAGpy/.snakemake/conda/377a289b
           /home/root/MAGpy/.snakemake/conda/8a832417
           base
        *  /root/miniconda3
```

I think the directive means snakemake (v5.11.2) was sourcing the
`.bashrc` when starting the shell for each rule and resetting to the
base environment before executing the rule script commands.

Without that directive the code everything seems to run fine for the
tests and full pipeline when tested with snakemake version >5